IKVPREL1L-1355: mmc: sdhci-msm: shutdown clock after probe
Propagated from IKVPREL2KK-10835

The sdhci-msm host driver leaves the clock running at 400kHz after probe
completes, but the sdhci platform driver assumes the clock is off.  When
the MMC core calls mmc_power_off(), the clocks are not actually disabled
until they are enabled once first.  In the case of a non-existant card,
that never happens and the clock is left running forever.

This commit shuts the clock down at the end of probe.

Change-Id: Ie47b9389cee0dcd94bc0f462f0a6b138d10bb064
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>