power: vm_bms: add support for battery aging compensation
The FCC (full charge capacity) and RBATT (battery resistance) gets
impacted based on the number of charge cycles.
Add a DT property which enables the battery-aging compensation for
FCC and RBATT based on charge-cycles. The DT property enables exposing
charge_cycle count to userspace and any non-zero value of this count
enables this feature in the userspace VBMS module

CRs-Fixed: 717939
Change-Id: I2230738d03684388db4302d2430187794fa26734
Signed-off-by: Ashay Jaiswal <ashayj@codeaurora.org>
Signed-off-by: Syukri Lajin <ginger_boy92@hotmail.com>