ASoC: wcd: correct the swap plug type detection
For swap plug type detection, only if it's detected as swap type
for four consecutive times, it will be set as swap. If there is
some problem during the detection, the longest mis-operation
period we can wait for is also four-consecutive-time detection.
So even if there is some unexpected behavior like HPH PA being
turned on, the fourth detection won't become the singular decisive
one.

CRs-Fixed: 806833

Change-Id: Icdb9d8e222b8cd9cb239deddfc637d6b874ad819
Signed-off-by: Walter Yang <yandongy@codeaurora.org>